+++ payment/src/main/java/com/ll/payment/model/dto/DepositWithdrawRequest.java
@@ -1,7 +1,7 @@
 package com.ll.payment.model.dto;
 
 public record DepositWithdrawRequest(
-        Long amount,
+        Integer amount,
         String referenceCode
 ) {
 }
